+++ packages/atomic/cypress/integration/facets/facet-selectors.ts
@@ -4,14 +4,16 @@ export const FacetSelectors = {
   facetStandard: 'atomic-facet',
   facetSearchbox: 'input[part="search-input"]',
   checkbox: 'input[type="checkbox"]',
+  label: 'label',
+  labelText: 'label span:nth-child(1)',
+  labelCount: 'label span:nth-child(2)',
   showMoreButton: 'button[part="show-more"]',
   showLessButton: 'button[part="show-less"]',
-  clearAllButton: 'button[part="reset-button"]',
+  clearAllButton: 'button[part="clear-button"]',
   numericFacet: 'atomic-numeric-facet',
   dateFacet: 'atomic-date-facet',
   categoryFacet: 'atomic-category-facet',
-  categoryFacetNextLevelButton: 'button',
-  categoryFacetClearLevelButton: 'div:nth-child(2) div:nth-child(1) button',
+  categoryFacetNextLevelButton: 'button[part="child"]',
 };
 
 export const BreadcrumbSelectors = {
@@ -21,8 +23,8 @@ export const BreadcrumbSelectors = {
 export const FacetAlias: IAlias = {
   facetShadow: '@facetShadow',
   facetUL: '@facetUL',
-  facetFirstValueLabel: '@facetFirstValueLabel',
-  facetSecondValueLabel: '@facetSecondValueLabel',
+  facetFirstValue: '@facetFirstValue',
+  facetSecondValue: '@facetSecondValue',
   facetAllValueLabel: '@facetAllValueLabel',
   facetAllValueCount: '@facetAllValueCount',
 };
@@ -38,11 +40,11 @@ const BreadcrumbAliasNoAtSign = aliasNoAtSignBuilder(BreadcrumbAlias);
 export function createBreadcrumbShadowAlias() {
   cy.get(BreadcrumbSelectors.breadcrumb)
     .shadow()
-    .find('.breadcrumbs')
+    .find('ul[part="breadcrumbs"]')
     .as(BreadcrumbAliasNoAtSign.breadcrumbFacet);
   cy.get(BreadcrumbSelectors.breadcrumb)
     .shadow()
-    .find('.breadcrumb-clear')
+    .find('button[part="breadcrumb-clear-all"]')
     .as(BreadcrumbAliasNoAtSign.breadcrumbClearAllFilter);
 }
 export function createAliasShadow(field: string, facetMainSelector?: string) {
@@ -64,10 +66,10 @@ export function createAliasFacetUL(field: string, facetMainSelector?: string) {
     .as(FacetAliasNoAtSign.facetUL);
   cy.get(FacetAlias.facetUL)
     .find('li:nth-child(1)')
-    .as(FacetAliasNoAtSign.facetFirstValueLabel);
+    .as(FacetAliasNoAtSign.facetFirstValue);
   cy.get(FacetAlias.facetUL)
     .find('li:nth-child(2)')
-    .as(FacetAliasNoAtSign.facetSecondValueLabel);
+    .as(FacetAliasNoAtSign.facetSecondValue);
 
   const facetValueLabelCSS =
     facetMainSelector === FacetSelectors.categoryFacet ? 'button' : 'label';