+++ packages/atomic/src/components/facets/facet-value/facet-value.tsx
@@ -1,39 +1,36 @@
-import {h} from '@stencil/core';
+import {h, FunctionalComponent} from '@stencil/core';
 import {randomID} from '../../../utils/utils';
 
 export type FacetValueProps = {
   facetValueSelected: VoidFunction;
   label: string;
-  numberOfResults: number;
+  numberOfResults: string;
   isSelected: boolean;
   ariaLabel: string;
 };
 
-export const FacetValue = (props: FacetValueProps) => {
-  const id = randomID('input');
+export const FacetValue: FunctionalComponent<FacetValueProps> = (props) => {
+  const id = randomID('facet-value-');
   return (
-    <li
-      part="facet-value"
-      class="flex flex-row items-center py-2 lg:py-1 cursor-pointer text-lg lg:text-base"
-      aria-label={props.ariaLabel}
-    >
+    <li part="value" class="flex flex-row items-center py-1">
       <input
         type="checkbox"
         checked={props.isSelected}
-        class="facet-value-checkbox w-5 h-5"
+        class="w-5 h-5 flex-none"
         id={id}
         name={id}
         onClick={() => props.facetValueSelected()}
+        aria-label={props.ariaLabel}
       />
       <label
         htmlFor={id}
-        class="ml-3 text-on-background cursor-pointer flex-shrink whitespace-nowrap overflow-ellipsis overflow-hidden flex flex-auto"
+        class="flex pl-3 text-on-background cursor-pointer ellipsed"
         title={props.label}
       >
-        <span class="whitespace-nowrap overflow-ellipsis overflow-hidden flex-auto">
+        <span part="value-label" class="ellipsed">
           {props.label}
         </span>
-        <span class="text-on-background-variant flex-shrink-0">
+        <span part="value-count" class="value-count">
           {props.numberOfResults}
         </span>
       </label>