+++ packages/headless/src/app/reducers.ts
@@ -34,7 +34,9 @@ export const pagination = paginationReducer;
 export const facetSet = facetSetReducer;
 export const facetSearchSet = specificFacetSearchSetReducer;
 export const search = searchReducer;
+
 export const dateFacetSet = dateFacetSetReducer;
+export const facetOrder = facetOrderReducer;
 export const numericFacetSet = numericFacetSetReducer;
 export const categoryFacetSet = categoryFacetSetReducer;
 export const facetOptions = facetOptionsReducer;
@@ -47,13 +49,11 @@ export const querySuggest = querySuggestReducer;
 export const sortCriteria = sortCriteriaReducer;
 export const context = contextReducer;
 export const didYouMean = didYouMeanReducer;
-export const resultPreview = resultPreviewReducer;
-
-export const debug = debugReducer;
-export const facetOrder = facetOrderReducer;
 export const fields = fieldsReducer;
 export const pipeline = pipelineReducer;
 export const searchHub = searchHubReducer;
+export const debug = debugReducer;
+export const resultPreview = resultPreviewReducer;
 export const version = versionReducer;
 
 export const history = undoable({